Add ReadAtPushRef method to git client
The ReadAtPushRef(ctx, branchName) method will invoke the following git
command:

`git rev-parse --abbrev-ref branchName@{push}`

The expected output is of the format `remoteName/branchName`

This should give us the proper RemoteRef that the branch pushes to. If it
fails, it is due to an unsupported git configuration. The errors this
emits are intended to help a user get out of this state.

The user should be able to navigate their way out of most broken state by
debugging the root `git rev-parse --abbrev-ref branchName@{push}` command.